Start moving classification feature functionality into the feature and out of the provider. Move most of the REST stuff from NLU into the Feature
