+++ includes/Classifai/Providers/Watson/Helpers.php
@@ -111,72 +111,6 @@ function get_classification_mode(): string {
 }
 
 /**
- * The list of post types that support classification.
- *
- * Defaults to 'post'.
- *
- * @return array
- */
-function get_supported_post_types(): array {
-	$feature    = new Classification();
-	$settings   = $feature->get_settings();
-	$post_types = [];
-
-	foreach ( $settings['post_types'] as $post_type => $enabled ) {
-		if ( ! empty( $enabled ) ) {
-			$post_types[] = $post_type;
-		}
-	}
-
-	/**
-	 * Filter post types supported for classification.
-	 *
-	 * @since 1.0.0
-	 * @hook classifai_post_types
-	 *
-	 * @param {array} $post_types Array of post types to be classified.
-	 *
-	 * @return {array} Array of post types.
-	 */
-	$post_types = apply_filters( 'classifai_post_types', $post_types );
-
-	return $post_types;
-}
-
-/**
- * The list of post statuses that support classification.
- *
- * Defaults to 'publish'.
- *
- * @return array
- */
-function get_supported_post_statuses(): array {
-	$feature       = new Classification();
-	$settings      = $feature->get_settings();
-	$post_statuses = [];
-
-	foreach ( $settings['post_statuses'] as $post_status => $enabled ) {
-		if ( ! empty( $enabled ) ) {
-			$post_statuses[] = $post_status;
-		}
-	}
-
-	/**
-	 * Filter post statuses supported for classification.
-	 *
-	 * @since 1.7.1
-	 * @hook classifai_post_statuses
-	 *
-	 * @param {array} $post_types Array of post statuses to be classified.
-	 *
-	 * @return {array} Array of post statuses.
-	 */
-	$post_statuses = apply_filters( 'classifai_post_statuses', $post_statuses );
-
-	return $post_statuses;
-}
-
-/**
  * Returns a bool based on whether the specified feature is enabled
  *
  * @param string $classify_by category,keyword,entity,concept
